Moved toast.js from lib to components
